+++ ompi/mca/fcoll/vulcan/fcoll_vulcan_file_read_all.c
@@ -221,7 +221,7 @@ int mca_fcoll_vulcan_file_read_all (struct ompio_file_t *fh,
         }
     }
 
-    result_counts = (int *) malloc (fh->f_num_aggrs * fh->f_procs_per_group * sizeof(int));
+    result_counts = (int *) malloc ((size_t)fh->f_num_aggrs * fh->f_procs_per_group * sizeof(int));
     if (NULL == result_counts) {
         ret = OMPI_ERR_OUT_OF_RESOURCE;
         goto exit;
@@ -441,7 +441,7 @@ int mca_fcoll_vulcan_file_read_all (struct ompio_file_t *fh,
 #endif
     }
 
-    reqs = (ompi_request_t **)malloc ((fh->f_procs_per_group + 1 )*fh->f_num_aggrs *sizeof(ompi_request_t *));
+    reqs = (ompi_request_t **)malloc (((size_t)fh->f_procs_per_group + 1 )*fh->f_num_aggrs *sizeof(ompi_request_t *));
     if (NULL == reqs) {
         opal_output (1, "OUT OF MEMORY\n");
         ret = OMPI_ERR_OUT_OF_RESOURCE;
@@ -520,7 +520,7 @@ int mca_fcoll_vulcan_file_read_all (struct ompio_file_t *fh,
 	end_read_time = ompi_wtime();
 	read_time += end_read_time - start_read_time;
 #endif
-	ret = ompi_request_wait_all ((fh->f_procs_per_group + 1 )*fh->f_num_aggrs,
+	ret = ompi_request_wait_all (((size_t)fh->f_procs_per_group + 1 )*fh->f_num_aggrs,
                                      reqs, MPI_STATUS_IGNORE);
         if (OMPI_SUCCESS != ret){
             goto exit;
@@ -542,7 +542,7 @@ int mca_fcoll_vulcan_file_read_all (struct ompio_file_t *fh,
                 goto exit;
             }
         }
-	ret = ompi_request_wait_all ((fh->f_procs_per_group + 1 )*fh->f_num_aggrs,
+	ret = ompi_request_wait_all (((size_t)fh->f_procs_per_group + 1 )*fh->f_num_aggrs,
                                      reqs, MPI_STATUS_IGNORE);
         if (OMPI_SUCCESS != ret){
             goto exit;
